+++ ChangeLog.md
@@ -2,6 +2,7 @@
 
 ## Unreleased
 * Server: Fix warning when opentelemetry is disabled
+* Bridge: Include the Kafka partition in consumer idempotency keys
 
 ## Version 1.98.0
 * CLI, Server, Bridge: Set OCI metadata on Docker images